+++ parser.py
@@ -1,4 +1,4 @@
-
+import numpy as np
 
 class FileParser(object):
     def __init__(self, numOfNodes, filePath):
@@ -16,7 +16,7 @@ class FileParser(object):
             line = line.split(" ")
             nodes_coord = []
             for idx in range(0, 2 * self.numOfNodes, 2):
-                nodes_coord.append([float(line[idx]), float(line[idx + 1])])
+                nodes_coord.append([np.float64(line[idx]), np.float64(line[idx + 1])])
             tour = [int(node) - 1 for node in line[line.index('output') + 1:-1]][:-1]
             self.coords.append(nodes_coord)
             self.tours.append(tour)
